refactor: 공용 컴포넌트를 Common 하위 폴더로 이동
Card, Button, Modal과 같이 공용으로 사용되는 컴포넌트를 Common 하위
폴더로  변경
