+++ frontend/src/components/DonationModal/DonationModal.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import Modal from '@/components/Modal';
-import Button from '@/components/Button';
+import Modal from '@/components/Common/Modal';
+import Button from '@/components/Common/Button';
 import DonationItemList from './DonationItemList';
 
 export default function DonationModal({ onClose, onSuccess }) {
