+++ frontend/src/components/Header/Header.jsx
@@ -6,7 +6,7 @@ import { Link } from 'react-router-dom';
 import HeaderLogo from '@/assets/images/header-logo.svg';
 import UserIcon from '@/assets/images/user-icon.svg';
 import { flexMixin } from '@/styles/mixins';
-import Button from '@/components/Button';
+import Button from '@/components/Common/Button';
 import LoginModal from '@/components/LoginModal';
 
 export default function Header({ isSigin }) {
