+++ frontend/src/components/LoginModal/LoginModal.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
-import Modal from '@/components/Modal';
-import Button from '@/components/Button';
+import Modal from '@/components/Common/Modal';
+import Button from '@/components/Common/Button';
 
 export default function LoginModal({ onClose }) {
     return (
